check-process: Select matching process to match what --help says
Currently -z, -r, -P, and -T *reject* matching processes. So there
is no way to check e.g. if there are processes that have more than
100 threads. -T 100 filters out such processes and reports that
there are other processes that didn't match the filter.

This patch fixes the issue by `select!`ing matching processes. For
the example above, we should be able to

  check-process.rb -T 100 --critical-over 1

to detect such processes.