Allow additional args to be passed to underlying tools
We need to send a particular arg to checkrestart to filter out some
false positives.

This change allows for any additional args to be passed through - rather
than restricting to the current known args. We have allowed it for
checkrestart and the needs-restarting tool. Though we don't know for
sure what options needs-restarting can take.

Appease Rubocop

Update the changelog