remove checks that do not actually make sense anymore
We previously checked if `sys-proctable` was greater than `0.9.8` when it was originally introduced in https://github.com/sensu-plugins/sensu-plugins-process-checks/commit/32044984dcb647ff63260435ebfd3d96894a26a0#diff-681f43ba752e5217d82b23a247fd5c79 it was `0.9.6` which would have fit this condition. We are currently requiring `0.9.8` which makes this check of no real value. We are now requiring `~> 1.1`.

On top of that I am not sure the logic actually was ever correct but I don't care enough to check if that is right as it is now irrelevant.

Signed-off-by: Ben Abrams <me@benabrams.it>